Unify the Base class for testing.
