+++ authentication/tests/test_promotions.py
@@ -1,20 +1,20 @@
 import json
 
 from associations.models import User
-from backend.tests_utils import BackendTestCase
+from backend.tests_utils import BaseTestCase
 
 
-class PromotionsTestCase(BackendTestCase):
+class PromotionsTestCase(BaseTestCase):
     fixtures = ('birthdays_test.json',)
 
     def test_if_not_logged_in_then_401(self):
-        res = self.get('promotions/')
+        res = self.get('/promotions/')
         self.assertStatusCode(res, 401)
 
     def test_promotions(self):
         self.user = self.create_and_login_user('17simple')
 
-        res = self.get('promotions/')
+        res = self.get('/promotions/')
         self.assertStatusCode(res, 200)
 
         promotions = json.loads(res.content)['promotions']
